FINALLY ADDED TOKEN FUNCTIONALITY

+++ videgameforum_backend/models/posts.model.go
@@ -9,7 +9,7 @@ type Post struct {
 	Post_Text  string
 	Post_Date  string
 	Post_ID    int
-	Users_ID   int
+	Users_ID   int64
 	Post_Title string
 	Forum_ID   int
 }
